Remove CNAME for gnubee-git version.
The CNAME file should only exist in the git tree
which actually displays gnubee.org.
In order to easily view development in the gnubee-git
project we have this branch which excludes that file
and can be viewed at
    https://gnubee-git.github.io/GnuBee_Docs/

Updates should go to the gh-pages-gnubee.org branch first,
then be merged into here.